refactored all source code.

+++ src/jp/ac/keio/sfc/ht/sox/tools/SaveAllNodeInformation.java
@@ -1,3 +1,9 @@
+/*
+ * This is a tool for saving all sensor node meta information.
+ * Usage:java SaveAllNodeInformation [directory name]
+ * Then, it saves all meta information into specified directory.
+ */
+
 package jp.ac.keio.sfc.ht.sox.tools;
 
 import java.io.File;
@@ -12,35 +18,36 @@ import org.simpleframework.xml.transform.Transform;
 import jp.ac.keio.sfc.ht.sox.protocol.Device;
 import jp.ac.keio.sfc.ht.sox.soxlib.SoxConnection;
 import jp.ac.keio.sfc.ht.sox.soxlib.SoxDevice;
+import jp.ac.keio.sfc.ht.sox.soxlib.SoxEnumTransform;
 
 public class SaveAllNodeInformation {
 
 	public static void main(String[] args) {
 		try {
-			new SaveAllNodeInformation();
+			new SaveAllNodeInformation(args[0]);
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
 	}
 
-	public SaveAllNodeInformation() throws Exception {
+	public SaveAllNodeInformation(String dir) throws Exception {
 
-		SoxConnection con = new SoxConnection("sox.ht.sfc.keio.ac.jp",
-				"takuro", "minatakuro", true);
+		SoxConnection con = new SoxConnection("sox.ht.sfc.keio.ac.jp", true);
 
 		List<String> list = con.getAllSensorList();
 		for (int i = 0; i < list.size(); i++) {
 			String s = list.get(i);
-			File file = new File("backup2/" + s + ".txt");
-			FileWriter filewriter = new FileWriter(file);
-			
-			if (!s.startsWith("タイムズ") ) {
+
+			File file = new File(dir + "/" + s + ".txt");
+			if (!file.exists()) {
+				FileWriter filewriter = new FileWriter(file);
+
 				System.out.println("getting meta data from " + s + " ..." + i
 						+ "/" + list.size());
 
 				SoxDevice dev = new SoxDevice(con, s);
 				Device device = dev.getDevice();
-			
+
 				// transform device object into XML string
 				if (device != null) {
 					StringWriter writer = new StringWriter();
@@ -68,8 +75,9 @@ public class SaveAllNodeInformation {
 					filewriter.write(writer.toString());
 
 				}
+
+				filewriter.close();
 			}
-			filewriter.close();
 		}
 
 		System.exit(0);
